Turn on strict mode for TS

+++ src/main/routes/helm-route.ts
@@ -17,7 +17,7 @@ class HelmApiRoute extends LensApi {
     try {
       const chart = await helmService.getChart(params.repo, params.chart, query.get("version"));
 
-      this.respondJson(response, chart);
+      this.respondJson(response, chart ?? {});
     } catch (error) {
       this.respondText(response, error, 422);
     }
@@ -29,7 +29,7 @@ class HelmApiRoute extends LensApi {
     try {
       const values = await helmService.getChartValues(params.repo, params.chart, query.get("version"));
 
-      this.respondJson(response, values);
+      this.respondJson(response, values ?? {});
     } catch (error) {
       this.respondText(response, error, 422);
     }
@@ -38,10 +38,16 @@ class HelmApiRoute extends LensApi {
   public async installChart(request: LensApiRequest) {
     const { payload, cluster, response } = request;
 
+    if (!cluster) {
+      logger.error("[HELM-ROUTE]: no cluster defined on installChart request");
+
+      return this.respondText(response, "No Cluster defined on request", 404);
+    }
+
     try {
       const result = await helmService.installChart(cluster, payload);
 
-      this.respondJson(response, result, 201);
+      this.respondJson(response, result ?? {}, 201);
     } catch (error) {
       logger.debug(error);
       this.respondText(response, error, 422);
@@ -51,10 +57,16 @@ class HelmApiRoute extends LensApi {
   public async updateRelease(request: LensApiRequest) {
     const { cluster, params, payload, response } = request;
 
+    if (!cluster) {
+      logger.error("[HELM-ROUTE]: no cluster defined on updateRelease request");
+
+      return this.respondText(response, "No Cluster defined on request", 404);
+    }
+
     try {
       const result = await helmService.updateRelease(cluster, params.release, params.namespace, payload );
 
-      this.respondJson(response, result);
+      this.respondJson(response, result ?? {});
     } catch (error) {
       logger.debug(error);
       this.respondText(response, error, 422);
@@ -64,10 +76,16 @@ class HelmApiRoute extends LensApi {
   public async rollbackRelease(request: LensApiRequest) {
     const { cluster, params, payload, response } = request;
 
+    if (!cluster) {
+      logger.error("[HELM-ROUTE]: no cluster defined on rollbackRelease request");
+
+      return this.respondText(response, "No Cluster defined on request", 404);
+    }
+
     try {
       const result = await helmService.rollback(cluster, params.release, params.namespace, payload.revision);
 
-      this.respondJson(response, result);
+      this.respondJson(response, result ?? {});
     } catch (error) {
       logger.debug(error);
       this.respondText(response, error, 422);
@@ -77,10 +95,16 @@ class HelmApiRoute extends LensApi {
   public async listReleases(request: LensApiRequest) {
     const { cluster, params, response } = request;
 
+    if (!cluster) {
+      logger.error("[HELM-ROUTE]: no cluster defined on listReleases request");
+
+      return this.respondText(response, "No Cluster defined on request", 404);
+    }
+
     try {
       const result = await helmService.listReleases(cluster, params.namespace);
 
-      this.respondJson(response, result);
+      this.respondJson(response, result ?? {});
     } catch(error) {
       logger.debug(error);
       this.respondText(response, error, 422);
@@ -89,6 +113,12 @@ class HelmApiRoute extends LensApi {
 
   public async getRelease(request: LensApiRequest) {
     const { cluster, params, response } = request;
+
+    if (!cluster) {
+      logger.error("[HELM-ROUTE]: no cluster defined on getRelease request");
+
+      return this.respondText(response, "No Cluster defined on request", 404);
+    }
 
     try {
       const result = await helmService.getRelease(cluster, params.release, params.namespace);
@@ -103,10 +133,16 @@ class HelmApiRoute extends LensApi {
   public async getReleaseValues(request: LensApiRequest) {
     const { cluster, params, response } = request;
 
+    if (!cluster) {
+      logger.error("[HELM-ROUTE]: no cluster defined on getReleaseValues request");
+
+      return this.respondText(response, "No Cluster defined on request", 404);
+    }
+
     try {
       const result = await helmService.getReleaseValues(cluster, params.release, params.namespace);
 
-      this.respondText(response, result);
+      this.respondText(response, result ?? "");
     } catch (error) {
       logger.debug(error);
       this.respondText(response, error, 422);
@@ -115,6 +151,12 @@ class HelmApiRoute extends LensApi {
 
   public async getReleaseHistory(request: LensApiRequest) {
     const { cluster, params, response } = request;
+
+    if (!cluster) {
+      logger.error("[HELM-ROUTE]: no cluster defined on getReleaseHistory request");
+
+      return this.respondText(response, "No Cluster defined on request", 404);
+    }
 
     try {
       const result = await helmService.getReleaseHistory(cluster, params.release, params.namespace);
@@ -129,10 +171,16 @@ class HelmApiRoute extends LensApi {
   public async deleteRelease(request: LensApiRequest) {
     const { cluster, params, response } = request;
 
+    if (!cluster) {
+      logger.error("[HELM-ROUTE]: no cluster defined on deleteRelease request");
+
+      return this.respondText(response, "No Cluster defined on request", 404);
+    }
+
     try {
       const result = await helmService.deleteRelease(cluster, params.release, params.namespace);
 
-      this.respondJson(response, result);
+      this.respondJson(response, result ?? "");
     } catch (error) {
       logger.debug(error);
       this.respondText(response, error, 422);
